BACK_END/requirements.txt
# AI Interview - Backend Dependencies (Python 3.13)

# Web framework & server
fastapi==0.111.1
uvicorn[standard]==0.30.3
starlette==0.37.2

# Configuration & multipart handling
python-dotenv==1.0.1
python-multipart==0.0.9

# Data validation
pydantic==2.8.2

# HTTP clients
requests==2.32.3
httpx==0.27.0

# Authentication & security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3

# OpenAI API client
openai==1.37.1

# AWS (per Polly TTS)
boto3==1.34.141

# MongoDB async
pymongo==4.8.0
motor==3.5.1

# Gestione file e multimedia
python-ffmpeg==2.0.11

# Elaborazione del testo e NLP
spacy==3.8.7
sentence-transformers==3.0.1
#model
spacy
https://github.com/explosion/spacy-models/releases/download/it_core_news_sm-3.8.0/it_core_news_sm-3.8.0-py3-none-any.whl


numpy==2.0.1
json5==0.9.25
python-docx==1.1.2
unidecode==1.3.8
rapidfuzz==3.9.4

# Elaborazione dati
pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1
xlsxwriter==3.2.0

# Tipizzazione
typing-extensions==4.12.2

